# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AnalyzeRiskResponse:
    properties:
      ai_risk_score:
        example: 35.5
        type: number
      flagged_clauses:
        items:
          $ref: '#/definitions/services.FlaggedClause'
        type: array
    type: object
...
  handlers.CreateCampaignResponse:
    properties:
      ai_target_segment_criteria:
        example: '{"criteria": "contacts with open rate > 30%"}'
        type: string
      campaign_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.CreateLeadResponse:
    properties:
      ai_win_probability:
        example: 65
        type: integer
      assigned_to:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      contact_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handlers.CreateTicketResponse:
    properties:
      ai_sentiment_score:
        example: -0.45
        type: number
      ai_suggested_response:
        example: Thank you for reaching out...
        type: string
      priority:
        example: high
        type: string
      ticket_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.ScheduleFieldVisitResponse:
    properties:
      estimated_travel_time_minutes:
        example: 25
        type: integer
      visit_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.analyzeRiskRequest:
    properties:
      contract_text:
        type: string
      quote_id:
        type: string
    type: object
...
  handlers.createCampaignRequest:
    properties:
      budget:
        type: number
      channel:
        type: string
      name:
        type: string
    type: object
...
  handlers.createLeadRequest:
    properties:
      company_name:
        type: string
      email:
        type: string
      estimated_deal_size:
        type: number
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  handlers.createTicketRequest:
    properties:
      contact_id:
        type: string
      description:
        type: string
      subject:
        type: string
    type: object
...
  handlers.launchCampaignResponse:
    properties:
      campaign_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      estimated_reach:
        example: 12500
        type: integer
      status:
        example: active
        type: string
    type: object
...
  handlers.scheduleFieldVisitRequest:
    properties:
      contact_id:
        type: string
      location_lat:
        type: number
      location_long:
        type: number
      notes:
        type: string
      sales_rep_id:
        type: string
      scheduled_at:
        type: string
    type: object
...
  services.FlaggedClause:
    properties:
      clause:
        type: string
      risk_level:
        description: low, medium, high, critical
        type: string
      suggested_fix:
        type: string
    type: object
...
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@strata.dev
    name: Strata Support
  description: ERP-CRM Hybrid API — direct competitor to Odoo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Strata API
  version: 0.1.0
paths:
...
  /api/v1/crm/campaigns:
    post:
      consumes:
      - application/json
      description: Creates a new marketing campaign with AI-generated target segment
        criteria. Requires `crm.campaigns.write` permission.
      operationId: createCampaign
      parameters:
      - description: Campaign payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateCampaignResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Create marketing campaign
      tags:
      - CRM
  /api/v1/crm/campaigns/{campaign_id}/launch:
    post:
      description: Launches a previously created campaign, transitioning it to active
        status and returning the estimated reach. Requires `crm.campaigns.write` permission.
      operationId: launchCampaign
      parameters:
      - description: Campaign ID
        in: path
        name: campaign_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.launchCampaignResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Launch marketing campaign
      tags:
      - CRM
  /api/v1/crm/field-visits:
    post:
      consumes:
      - application/json
      description: Schedules an in-person field sales visit for a contact, returning
        the estimated travel time. Requires `crm.fieldvisits.write` permission.
      operationId: scheduleFieldVisit
      parameters:
      - description: Field visit payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.scheduleFieldVisitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ScheduleFieldVisitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Schedule field sales visit
      tags:
      - CRM
  /api/v1/crm/leads:
    post:
      consumes:
      - application/json
      description: Creates a new CRM contact as a lead, runs AI win probability scoring,
        and creates a linked deal. Requires `crm.leads.write` permission.
      operationId: createLead
      parameters:
      - description: Lead payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createLeadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateLeadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Create lead & trigger AI scoring
      tags:
      - CRM
  /api/v1/crm/quotes/risk-analysis:
    post:
      consumes:
      - application/json
      description: Runs AI risk analysis on a quote's contract text, identifying flagged
        clauses with risk levels and suggested fixes. Requires `crm.quotes.write`
        permission.
      operationId: analyzeContractRisk
      parameters:
      - description: Risk analysis payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.analyzeRiskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AnalyzeRiskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Analyze contract risk
      tags:
      - CRM
  /api/v1/crm/tickets:
    post:
      consumes:
      - application/json
      description: |-
        Creates a helpdesk ticket, runs AI sentiment analysis on the description, auto-assigns priority and routing. Requires `crm.tickets.write` permission.
        Creates a helpdesk ticket with AI sentiment analysis, auto-assigned priority, and an AI-generated suggested response. Requires `crm.tickets.write` permission.
      operationId: createTicket
      parameters:
      - description: Ticket payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateTicketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Create support ticket & auto-route
      tags:
      - CRM
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access (e.g. fleet telematics ingestion)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...